refactor(#608): split tests/integration.rs into a tests/integration/ module tree
Mechanical split of the 6,908-line single-file integration suite into one
module per CLI domain, following the seam map in audit document
019eb3b9-6694-7e42-945f-f90e1c0774db:

- tests/integration/main.rs holds only the mod declarations; the directory
  compiles as a single test binary so CARGO_BIN_EXE_legion and the shared
  helpers link once.
- tests/integration/common.rs owns the spawn-and-assert plumbing:
  legion_cmd, legion_cmd_with_home, assert_uuid_format (now a real
  uuid::Uuid::parse_str + version-7 check instead of the len/hyphen
  heuristic), seed_rate_limit_sample (shared by mesh and governance), and
  the new run_ok / run_ok_stderr / run_fail / run_with_stdin helpers.
- The run-and-assert idiom (~270 sites of output().unwrap() +
  assert!(status.success(), "...", from_utf8_lossy(stderr))) is converted
  to the helpers during the move, per domain file. Sites that are
  deliberately status-agnostic (flag-parse smoke tests) or assert on both
  streams of one successful run keep their bespoke shape.
- mcp_push_bridge_delivers_cross_process_post moved whole and untouched,
  including its load-bearing schema-warmup and stderr-drain comments.
- The piped-stdin spawn dance is collapsed into run_with_stdin at its five
  collapsible sites; the two bespoke MCP stream sites stay hand-rolled.

Test count before: 163. Test count after: 163. All green; clippy and fmt
clean.
